# Use list comprehension in audformat.utils.intersect()