--- a/tests/test_utils.py
+++ b/tests/test_utils.py
@@ -1038,7 +1038,7 @@ def test_index_has_overlap(obj, expected):
         (
             [
                 pd.Index([1, np.nan], dtype='Int64', name='idx'),
-                pd.Index([1, 2, 3], name='idx'),
+                pd.Index([1, 2, 3], dtype='Int64', name='idx'),
             ],
             pd.Index([1], dtype='Int64', name='idx'),
         ),
@@ -1084,10 +1084,10 @@ def test_index_has_overlap(obj, expected):
 )
 def test_intersect(objs, expected):
     for permuted_objs in itertools.permutations(objs):
-        pd.testing.assert_index_equal(
-            audformat.utils.intersect(permuted_objs),
-            expected,
-        )
+        index = audformat.utils.intersect(permuted_objs)
+        if len(index) > 0:
+            index = index.sort_values()
+        pd.testing.assert_index_equal(index, expected)
     # Ensure A ∩ (B ∩ C) == (A ∩ B) ∩ C
     if len(objs) > 2:
         pd.testing.assert_index_equal(