Addresses half of https://github.com/audeering/audformat/issues/245

This implements `audformat.utils.intersect()` by converting indices to lists before checking for intersections,
and creating a new index at the end of the operation.
It also suggest to not sort the objects, in order to allow the test proposed in https://github.com/audeering/audformat/pull/254 to pass and preserve original index orders (output is no longer commutative then).

I benchmarked the current code using `list()` against the implementation of the master. The code for the benchmark is given below.


**Results for last obj of same size as first object**

| Table type | Index samples | v0.14.3 | master | list+sort | list-sort | pull request |
| --- | --- | --- | --- | --- | --- | --- |
| filewise | (100000, 500000, 100000) | 0.16 s | 0.19 s  | 0.14 s | 0.16 s | 0.16 s |
| filewise | (100000, 500000, 10000) | 0.12 s | 0.12 s  | 0.07 s | 0.14 s | 0.07 s |
| segmented | (100000, 500000, 100000) | 6.04 s | 14.51 s | 2.81 s | 3.17 s | 2.71 s |
| segmented | (100000, 500000, 10000) | 5.14 s | 12.91 s | 0.24 s | 3.04 s | 0.27 s |

![image](https://user-images.githubusercontent.com/173624/182303889-afcb4d6d-7c3e-40ca-89b5-70f0ed82c3af.png)

---

Code for benchmark:

```python
import audformat
import numpy as np
import pandas as pd
import random
import time


random.seed(0)
samples = 100000


def main():
    nsamples = 100000
    repetitions = 10

    for description, samples in [
            (
                'same length',
                [nsamples, 5 * nsamples, nsamples],
            ),
            (
                '1/10 length',
                [nsamples, 5 * nsamples, int(nsamples / 10)],
            ),
    ]:  
        for index in [filewise, segmented]:
            times = []
            for _ in range(repetitions):
                objs = index(samples)
                times.append(measure(objs))
            print(
                f'Execution time {index.__name__} {description}: '
                f'{np.mean(times):.2f}s'
            )


def measure(objs):
    start = time.time()
    _ = audformat.utils.intersect(objs)
    end = time.time()
    return end - start


def filewise(samples):
    index1 = audformat.filewise_index([f"f{n}" for n in range(samples[0])])
    index2 = audformat.filewise_index([f"f{n}" for n in range(samples[1])])
    idx = [n for n in range(len(index2))]
    index3 = index2[random.sample(idx, k=samples[2])]
    return [index1, index2, index3]


def segmented(samples):
    index1 = audformat.segmented_index(
        [f"f{n}" for n in range(samples[0])],
        random.choices([0, 1, 2, 3], k=samples[0]),
        random.choices([4, 5, 6, pd.NaT], k=samples[0]),
    )
    index2 = audformat.segmented_index(
        [f"f{n}" for n in range(samples[1])],
        random.choices([0, 1, 2, 3], k=samples[1]),
        random.choices([4, 5, 6, pd.NaT], k=samples[1]),
    )
    idx = [n for n in range(len(index2))]
    index3 = index2[random.sample(idx, k=samples[2])]
    return [index1, index2, index3]


if __name__ == '__main__':
    main()
```