@hagenw 2022-07-31
Tests are failing with the new `flake8` v5.0.0 release from https://github.com/PyCQA/flake8/releases/tag/5.0.0