@hagenw 2022-08-01
https://github.com/audeering/audformat/pull/257 needs to be merged first to fix the test errors.