@hagenw 2022-08-01
> > Table type 	Index entries 	master 	set 	list+sort 	list-sort
> > filewise 	1,000,000 	0.82 s 	2.78 s 	0.11 s 	1.14 s
> > segmented 	100,000 	7.81 s 	0.24 s 	0.08 s 	1.98 s
> 
> So list with sorting is faster than list without sorting? Isn't that strange?

It's not, because with intersection you can gain speed in the for loop by picking the smallest index as the starting index to compare against. This you can no longer do without sorting (or if we still want to use it we would need a method to reverse the sorting at the end).