@frankenjoe 2022-08-01
> It's not, because with intersection you can gain speed in the for loop by picking the smallest index as the starting index to compare against. This you can no longer do without sorting (or if we still want to use it we would need a method to reverse the sorting at the end).

Ah, now I get it. We are sorting only by length of the objects. That makes sense of course.