@frankenjoe 2022-08-01
If we stick with the supposed solution of using list comprehension, it might still be faster to use `set()` within the comprehension:

```python
index = [idx for idx in index if idx in set(obj)]
```