@hagenw 2022-08-01
> If we stick with the proposed solution of using list comprehension, it might still be faster to use `set()` within the comprehension:
> 
> ```python
> obj = set(obj)
> index = [idx for idx in index if idx in obj]
> ```

On my machine I was not able to finish the runtime test for your suggestion. Maybe it needs too much memory?
I will do another check if `timeit` is the problem or using `set()` is the problem.