@hagenw 2022-08-01
Using the code mentioned above we get:

filewise: 3.10 s
segmented: 0.27 s

For whatever reason `set()` struggles with our filewise indices. I saw the same result when I directly used `set().intersection()` for the implementation.