@frankenjoe 2022-08-01
Yes, that would explain why we are so much slower for `filewise` with `set()`. Just that we now it's indeed the size of the input could you please repeat with `100.000` in both cases?