@hagenw 2022-08-01
> Yes, that would explain why we are so much slower for `filewise` with `set()`. Just that we now it's indeed the size of the input could you please repeat with `100.000` entries in both cases?

If we use 100.000 entries in both cases we get the same result:
* filewise: 0.25 s
* segmented: 0.26 s