@frankenjoe 2022-08-01
In fact, I wonder why `set()` can have any benefit, since we know that indices in `audformat` cannot have duplicated items by design.