@hagenw 2022-08-01
> In fact, I wonder why `set()` can bring any benefit, since we know that indices in `audformat` cannot have duplicated items by design.

You are right, the length of the `set(obj)` is identical to the length of `obj`.
The difference is an artifact of the repeated measurement time as sets seem to be different cached in Python.
When only focusing on the segmented indices we get **with set**:
* first execution time: 5.80 s
* second execution time: 0.22 s

and **without set**:
* first execution time: 3.05 s
* second execution time: 1.99 s

As repeated execution is not really a feature we target here, I guess we are better without using set.
But maybe I should revisit the benchmarks I did and focus on single execution time?