@frankenjoe 2022-08-01
> But maybe I should revisit the benchmarks I did and focus on single execution time?

Yes, I think that makes sense.