@hagenw 2022-08-02
> > But maybe I should revisit the benchmarks I did and focus on single execution time?
> 
> Yes, I think that makes sense.

I updated the benchmark code and benchmark results. I also included now v0.14.3 for comparison.

In addition, I found a way to still use sort, but restore the order at the end.