@frankenjoe 2022-08-02
I guess you plan to add a test that `Table.pick_index()` preserves order in https://github.com/audeering/audformat/pull/254, right?