@hagenw 2022-08-02
> I guess you plan to add a test that `Table.pick_index()` preserves order in #254, right?

Yes, or would you prefer to include it here already?