@frankenjoe 2022-08-02
> Yes, or would you prefer to include it here already?

No let's do it in #254, maybe after also optimizing `symmetric_differences()` using lists.